+++ python/sglang/srt/layers/attention/triton_ops/decode_attention.py
@@ -231,8 +231,8 @@ def _decode_att_m_fwd(
     q,
     k_buffer,
     v_buffer,
-    k_scale_zeros_buffer,
-    v_scale_zeros_buffer,
+    k_scales_zeros_buffer,
+    v_scales_zeros_buffer,
     att_out,
     Req_to_tokens,
     B_req_idx,
@@ -264,22 +264,22 @@ def _decode_att_m_fwd(
     BLOCK_DV = triton.next_power_of_2(Lv)
 
     if USE_INT8_KV:
-        k_scale_zeros_stride0 = k_scale_zeros_buffer.stride(0)
-        k_scale_zeros_stride1 = k_scale_zeros_buffer.stride(1)
-        v_scale_zeros_stride0 = v_scale_zeros_buffer.stride(0)
-        v_scale_zeros_stride1 = v_scale_zeros_buffer.stride(1)
+        k_scales_zeros_stride0 = k_scales_zeros_buffer.stride(0)
+        k_scales_zeros_stride1 = k_scales_zeros_buffer.stride(1)
+        v_scales_zeros_stride0 = v_scales_zeros_buffer.stride(0)
+        v_scales_zeros_stride1 = v_scales_zeros_buffer.stride(1)
     else:
-        k_scale_zeros_stride0 = None
-        k_scale_zeros_stride1 = None
-        v_scale_zeros_stride0 = None
-        v_scale_zeros_stride1 = None
+        k_scales_zeros_stride0 = None
+        k_scales_zeros_stride1 = None
+        v_scales_zeros_stride0 = None
+        v_scales_zeros_stride1 = None
 
     _fwd_kernel_stage1[grid](
         q,
         k_buffer,
         v_buffer,
-        k_scale_zeros_buffer,
-        v_scale_zeros_buffer,
+        k_scales_zeros_buffer,
+        v_scales_zeros_buffer,
         sm_scale,
         Req_to_tokens,
         B_req_idx,
@@ -290,12 +290,12 @@ def _decode_att_m_fwd(
         q.stride(1),
         k_buffer.stride(0),
         k_buffer.stride(1),
-        k_scale_zeros_stride0,
-        k_scale_zeros_stride1,
+        k_scales_zeros_stride0,
+        k_scales_zeros_stride1,
         v_buffer.stride(0),
         v_buffer.stride(1),
-        v_scale_zeros_stride0,
-        v_scale_zeros_stride1,
+        v_scales_zeros_stride0,
+        v_scales_zeros_stride1,
         att_out.stride(0),
         att_out.stride(1),
         att_out.stride(2),
@@ -536,8 +536,8 @@ def _decode_grouped_att_m_fwd(
     q,
     k_buffer,
     v_buffer,
-    k_scale_zeros_buffer,
-    v_scale_zeros_buffer,
+    k_scales_zeros_buffer,
+    v_scales_zeros_buffer,
     att_out,
     Req_to_tokens,
     B_req_idx,
@@ -587,22 +587,22 @@ def _decode_grouped_att_m_fwd(
         extra_kargs = {"waves_per_eu": 4, "matrix_instr_nonkdim": 16, "kpack": 2}
 
     if USE_INT8_KV:
-        k_scale_zeros_stride0 = k_scale_zeros_buffer.stride(0)
-        k_scale_zeros_stride1 = k_scale_zeros_buffer.stride(1)
-        v_scale_zeros_stride0 = v_scale_zeros_buffer.stride(0)
-        v_scale_zeros_stride1 = v_scale_zeros_buffer.stride(1)
+        k_scales_zeros_stride0 = k_scales_zeros_buffer.stride(0)
+        k_scales_zeros_stride1 = k_scales_zeros_buffer.stride(1)
+        v_scales_zeros_stride0 = v_scales_zeros_buffer.stride(0)
+        v_scales_zeros_stride1 = v_scales_zeros_buffer.stride(1)
     else:
-        k_scale_zeros_stride0 = None
-        k_scale_zeros_stride1 = None
-        v_scale_zeros_stride0 = None
-        v_scale_zeros_stride1 = None
+        k_scales_zeros_stride0 = None
+        k_scales_zeros_stride1 = None
+        v_scales_zeros_stride0 = None
+        v_scales_zeros_stride1 = None
 
     _fwd_grouped_kernel_stage1[grid](
         q,
         k_buffer,
         v_buffer,
-        k_scale_zeros_buffer,
-        v_scale_zeros_buffer,
+        k_scales_zeros_buffer,
+        v_scales_zeros_buffer,
         sm_scale,
         Req_to_tokens,
         B_req_idx,
@@ -613,12 +613,12 @@ def _decode_grouped_att_m_fwd(
         q.stride(1),
         k_buffer.stride(0),
         k_buffer.stride(1),
-        k_scale_zeros_stride0,
-        k_scale_zeros_stride1,
+        k_scales_zeros_stride0,
+        k_scales_zeros_stride1,
         v_buffer.stride(0),
         v_buffer.stride(1),
-        v_scale_zeros_stride0,
-        v_scale_zeros_stride1,
+        v_scales_zeros_stride0,
+        v_scales_zeros_stride1,
         att_out.stride(0),
         att_out.stride(1),
         att_out.stride(2),
@@ -747,16 +747,16 @@ def decode_attention_fwd_normal(
     num_kv_splits,
     sm_scale,
     logit_cap=0.0,
-    k_scale_zeros_buffer=None,
-    v_scale_zeros_buffer=None,
+    k_scales_zeros_buffer=None,
+    v_scales_zeros_buffer=None,
     kv_cache_dtype=None,
 ):
     _decode_att_m_fwd(
         q,
         k_buffer,
         v_buffer,
-        k_scale_zeros_buffer,
-        v_scale_zeros_buffer,
+        k_scales_zeros_buffer,
+        v_scales_zeros_buffer,
         attn_logits,
         req_to_token,
         b_req_idx,
@@ -781,16 +781,16 @@ def decode_attention_fwd_grouped(
     num_kv_splits,
     sm_scale,
     logit_cap=0.0,
-    k_scale_zeros_buffer=None,
-    v_scale_zeros_buffer=None,
+    k_scales_zeros_buffer=None,
+    v_scales_zeros_buffer=None,
     kv_cache_dtype=None,
 ):
     _decode_grouped_att_m_fwd(
         q,
         k_buffer,
         v_buffer,
-        k_scale_zeros_buffer,
-        v_scale_zeros_buffer,
+        k_scales_zeros_buffer,
+        v_scales_zeros_buffer,
         attn_logits,
         req_to_token,
         b_req_idx,
@@ -815,8 +815,8 @@ def decode_attention_fwd(
     num_kv_splits,
     sm_scale,
     logit_cap=0.0,
-    k_scale_zeros_buffer=None,
-    v_scale_zeros_buffer=None,
+    k_scales_zeros_buffer=None,
+    v_scales_zeros_buffer=None,
     kv_cache_dtype=None,
 ):
     assert num_kv_splits == attn_logits.shape[2]
@@ -836,8 +836,8 @@ def decode_attention_fwd(
             num_kv_splits,
             sm_scale,
             logit_cap,
-            k_scale_zeros_buffer,
-            v_scale_zeros_buffer,
+            k_scales_zeros_buffer,
+            v_scales_zeros_buffer,
             kv_cache_dtype,
         )
     else:
@@ -854,172 +854,7 @@ def decode_attention_fwd(
             num_kv_splits,
             sm_scale,
             logit_cap,
-            k_scale_zeros_buffer,
-            v_scale_zeros_buffer,
+            k_scales_zeros_buffer,
+            v_scales_zeros_buffer,
             kv_cache_dtype,
         )
-
-
-@triton.jit
-def quantize_and_store(
-    cur_index,
-    data_ptr,
-    cache_ptr,
-    scale_zeros_ptr,
-    stride_bs,
-    stride_h,
-    stride_d,
-    stride_c_bs,
-    stride_c_h,
-    stride_c_d,
-    dest_index,
-    offs_h,
-    offs_d,
-    head_num,
-    szd_off,
-):
-    data = tl.load(
-        data_ptr
-        + cur_index * stride_bs
-        + offs_h[:, None] * stride_h
-        + offs_d[None, :] * stride_d,
-        mask=offs_h[:, None] < head_num,
-        other=0.0,
-    )
-
-    quant, scales, zeros = _quant_int8(data)
-    o_ptrs = (
-        cache_ptr
-        + dest_index * stride_bs
-        + offs_h[:, None] * stride_h
-        + offs_d[None, :] * stride_d
-    )
-    sz_ptrs_k = (
-        scale_zeros_ptr
-        + dest_index * stride_c_bs
-        + stride_c_h * offs_h[:, None] * stride_c_d
-    )
-    tl.store(o_ptrs, quant, mask=offs_h[:, None] < head_num)
-    tl.store(
-        sz_ptrs_k + szd_off[None, :] * 1,
-        scales[:, None],
-        mask=(offs_h[:, None] < head_num) & (szd_off[None, :] < 1),
-    )
-    tl.store(
-        sz_ptrs_k + szd_off[None, :] * 1,
-        zeros[:, None],
-        mask=(offs_h[:, None] < head_num) & (szd_off[None, :] == 1),
-    )
-
-
-@triton.jit
-def _fwd_kernel_quantize_cache_kv(
-    K_Status,
-    V_Status,
-    Dest_Idx,
-    K_Cache,
-    V_Cache,
-    K_Scale_Zeros,
-    V_Scale_Zeros,
-    stride_k_bs,
-    stride_k_h,
-    stride_k_d,
-    stride_v_bs,
-    stride_v_h,
-    stride_v_d,
-    stride_kv_sz_bs,
-    stride_kv_sz_h,
-    stride_kv_sz_d,
-    head_num,
-    BLOCK_DMODEL: tl.constexpr,
-    BLOCK_HEAD: tl.constexpr,
-):
-    cur_index = tl.program_id(0)
-    offs_h = tl.arange(0, BLOCK_HEAD)
-    offs_d = tl.arange(0, BLOCK_DMODEL)
-    dest_index = tl.load(Dest_Idx + cur_index)
-    szd_off = tl.arange(0, 2)
-
-    # Process K
-    quantize_and_store(
-        cur_index,
-        K_Status,
-        K_Cache,
-        K_Scale_Zeros,
-        stride_k_bs,
-        stride_k_h,
-        stride_k_d,
-        stride_kv_sz_bs,
-        stride_kv_sz_h,
-        stride_kv_sz_d,
-        dest_index,
-        offs_h,
-        offs_d,
-        head_num,
-        szd_off,
-    )
-
-    # Process V
-    quantize_and_store(
-        cur_index,
-        V_Status,
-        V_Cache,
-        V_Scale_Zeros,
-        stride_v_bs,
-        stride_v_h,
-        stride_v_d,
-        stride_kv_sz_bs,
-        stride_kv_sz_h,
-        stride_kv_sz_d,
-        dest_index,
-        offs_h,
-        offs_d,
-        head_num,
-        szd_off,
-    )
-
-
-def quantize_cache_kv(
-    k,
-    v,
-    dest_idx,
-    k_quantized_out,
-    k_scales_zeros,
-    v_quantized_out,
-    v_scales_zeros,
-):
-    bs = dest_idx.shape[0]
-    k_head_num = k.shape[1]
-    k_head_dim = k.shape[2]
-    assert (
-        k.shape[1] == k_quantized_out.shape[1]
-        and k.shape[2] == k_quantized_out.shape[2]
-    )
-    BLOCK_HEAD = triton.next_power_of_2(k_head_num)
-    grid = (bs,)
-    num_warps = min(max(k_head_dim // 256, 1), 8)
-
-    _fwd_kernel_quantize_cache_kv[grid](
-        k,
-        v,
-        dest_idx,
-        k_quantized_out,
-        v_quantized_out,
-        k_scales_zeros,
-        v_scales_zeros,
-        k.stride(0),
-        k.stride(1),
-        k.stride(2),
-        v.stride(0),
-        v.stride(1),
-        v.stride(2),
-        k_scales_zeros.stride(0),
-        k_scales_zeros.stride(1),
-        k_scales_zeros.stride(2),
-        k_head_num,
-        BLOCK_DMODEL=k_head_dim,
-        BLOCK_HEAD=BLOCK_HEAD,
-        num_warps=num_warps,
-        num_stages=1,
-    )
-    return
